Style room header icons for when they are toggled
